# console-chat/openapi.json
{"openapi":"3.1.0","info":{"title":"Console Chat API","version":"0.1.0"},"paths":{"/users/online":{"get":{"tags":["users"],"summary":"Online","description":"Set Status to online/Create Auth Token for temporary user","operationId":"online_users_online_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"username","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnlineResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/login":{"post":{"tags":["users"],"summary":"Login","description":"Login as permanent user","operationId":"login_users_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnlineResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/register":{"post":{"tags":["users"],"summary":"Register","description":"Register as permanent user","operationId":"register_users_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnlineResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/users/me":{"get":{"tags":["users"],"summary":"Get Me","operationId":"get_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPrivate"}}}}},"security":[{"HTTPBearer":[]}]}},"/rooms/":{"get":{"tags":["rooms"],"summary":"List Rooms","operationId":"list_rooms_rooms__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/StaticRoomPublic"},"type":"array","title":"Response List Rooms Rooms  Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/rooms/mine":{"get":{"tags":["rooms"],"summary":"Get My Rooms","operationId":"get_my_rooms_rooms_mine_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/StaticRoomPublic"},"type":"array","title":"Response Get My Rooms Rooms Mine Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/rooms/{room}":{"put":{"tags":["rooms"],"summary":"Create Room","operationId":"create_room_rooms__room__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"room","in":"path","required":true,"schema":{"type":"string","title":"Room"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoom"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["rooms"],"summary":"Update Room","operationId":"update_room_rooms__room__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"room","in":"path","required":true,"schema":{"type":"string","title":"Room"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoom"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["rooms"],"summary":"Delete Room","operationId":"delete_room_rooms__room__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"room","in":"path","required":true,"schema":{"type":"string","title":"Room"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["rooms"],"summary":"Get Room","operationId":"get_room_rooms__room__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"room","in":"path","required":true,"schema":{"type":"string","title":"Room"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/room/{room}":{"post":{"tags":["deprecated"],"summary":"Send","operationId":"send_room__room__post","deprecated":true,"security":[{"HTTPBearer":[]}],"parameters":[{"name":"room","in":"path","required":true,"schema":{"type":"string","title":"Room"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageSend"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["deprecated"],"summary":"Listen","operationId":"listen_room__room__get","deprecated":true,"security":[{"HTTPBearer":[]}],"parameters":[{"name":"room","in":"path","required":true,"schema":{"type":"string","title":"Room"}},{"name":"listen_seconds","in":"query","required":false,"schema":{"type":"integer","description":"How long to listen in seconds","default":30,"title":"Listen Seconds"},"description":"How long to listen in seconds"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AppearancePublic":{"properties":{"color":{"type":"string","maxLength":7,"minLength":7,"title":"Color"}},"type":"object","required":["color"],"title":"AppearancePublic"},"CreateRoom":{"properties":{"private_level":{"$ref":"#/components/schemas/RoomLevel"},"invite":{"anyOf":[{"items":{"anyOf":[{"type":"string","format":"uuid"},{"type":"string"}]},"type":"array"},{"type":"null"}],"title":"Invite"},"key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key"}},"type":"object","required":["private_level"],"title":"CreateRoom"},"Encrypted":{"properties":{"content_base64":{"type":"string","title":"Content Base64"},"nonce":{"type":"string","title":"Nonce"}},"type":"object","required":["content_base64","nonce"],"title":"Encrypted"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Json":{"anyOf":[{"additionalProperties":{"$ref":"#/components/schemas/Json"},"type":"object"},{"items":{"$ref":"#/components/schemas/Json"},"type":"array"},{"type":"string"},{"type":"integer"},{"type":"number"},{"type":"boolean"},{"type":"null"}]},"KeyRequest":{"properties":{"public_key":{"type":"string","title":"Public Key"}},"type":"object","required":["public_key"],"title":"KeyRequest"},"KeyResponse":{"properties":{"encrypted_symmetric_key":{"type":"string","title":"Encrypted Symmetric Key"},"check_msg":{"type":"string","title":"Check Msg"},"sender_public_key":{"type":"string","title":"Sender Public Key"}},"type":"object","required":["encrypted_symmetric_key","check_msg","sender_public_key"],"title":"KeyResponse"},"LoginData":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginData"},"MessageSend":{"properties":{"type":{"$ref":"#/components/schemas/MessageType","default":"PLAINTEXT"},"content":{"anyOf":[{"$ref":"#/components/schemas/Encrypted"},{"$ref":"#/components/schemas/Plaintext"},{"$ref":"#/components/schemas/KeyRequest"},{"$ref":"#/components/schemas/KeyResponse"},{"$ref":"#/components/schemas/SystemMessage"},{"type":"null"}],"title":"Content"},"send_at":{"type":"string","format":"date-time","title":"Send At"},"data":{"anyOf":[{"$ref":"#/components/schemas/Json"},{"type":"null"}]}},"type":"object","title":"MessageSend"},"MessageType":{"type":"string","enum":["ENCRYPTED","PLAINTEXT","KEY_REQUEST","KEY_RESPONSE","SYSTEM","JOIN","LEAVE"],"title":"MessageType"},"OnlineResponse":{"properties":{"token":{"$ref":"#/components/schemas/Token"},"user":{"type":"string","format":"uuid","title":"User"}},"type":"object","required":["token","user"],"title":"OnlineResponse"},"Plaintext":{"properties":{"content":{"type":"string","title":"Content"}},"type":"object","required":["content"],"title":"Plaintext"},"RegisterData":{"properties":{"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["password"],"title":"RegisterData"},"RoomLevel":{"type":"string","enum":["FREE","KEY","INVITE-ONLY","INVITE-AND-KEY"],"title":"RoomLevel"},"StaticRoomPublic":{"properties":{"name":{"type":"string","title":"Name"},"key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key"},"id":{"type":"integer","title":"Id"},"owner":{"$ref":"#/components/schemas/UserPublic"},"users":{"items":{"$ref":"#/components/schemas/UserPublic"},"type":"array","title":"Users"},"level":{"$ref":"#/components/schemas/RoomLevel"}},"type":"object","required":["name","id","owner","users","level"],"title":"StaticRoomPublic"},"SystemMessage":{"properties":{"content":{"type":"string","title":"Content"},"online_users":{"type":"integer","title":"Online Users"}},"type":"object","required":["content","online_users"],"title":"SystemMessage"},"Token":{"properties":{"token":{"type":"string","title":"Token"},"ttl":{"type":"integer","title":"Ttl"}},"type":"object","required":["token","ttl"],"title":"Token"},"UpdateRoom":{"properties":{"private_level":{"anyOf":[{"$ref":"#/components/schemas/RoomLevel"},{"type":"null"}]},"invite":{"anyOf":[{"items":{"anyOf":[{"type":"string","format":"uuid"},{"type":"string"}]},"type":"array"},{"type":"null"}],"title":"Invite"},"key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key"}},"type":"object","title":"UpdateRoom"},"UserPrivate":{"properties":{"username":{"type":"string","maxLength":100,"title":"Username","default":"anonym"},"user_type":{"$ref":"#/components/schemas/UserType","default":"GUEST"},"appearance":{"$ref":"#/components/schemas/AppearancePublic"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["appearance","id"],"title":"UserPrivate"},"UserPublic":{"properties":{"username":{"type":"string","maxLength":100,"title":"Username","default":"anonym"},"user_type":{"$ref":"#/components/schemas/UserType","default":"GUEST"},"appearance":{"$ref":"#/components/schemas/AppearancePublic"}},"type":"object","required":["appearance"],"title":"UserPublic"},"UserType":{"type":"string","enum":["GUEST","PERMANENT"],"title":"UserType"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}