# console-chat/openapi/src/models/encrypted.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct Encrypted {
    #[serde(rename = "content_base64")]
    pub content_base64: String,
    #[serde(rename = "nonce")]
    pub nonce: String,
}

impl Encrypted {
    pub fn new(content_base64: String, nonce: String) -> Encrypted {
        Encrypted {
            content_base64,
            nonce,
        }
    }
}