# console-chat/openapi/src/models/key_request.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct KeyRequest {
    #[serde(rename = "public_key")]
    pub public_key: String,
}

impl KeyRequest {
    pub fn new(public_key: String) -> KeyRequest {
        KeyRequest {
            public_key,
        }
    }
}