# console-chat/openapi/src/models/user_public.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct UserPublic {
    #[serde(rename = "username", skip_serializing_if = "Option::is_none")]
    pub username: Option<String>,
    #[serde(rename = "user_type", skip_serializing_if = "Option::is_none")]
    pub user_type: Option<models::UserType>,
    #[serde(rename = "appearance")]
    pub appearance: Box<models::AppearancePublic>,
}

impl UserPublic {
    pub fn new(appearance: models::AppearancePublic) -> UserPublic {
        UserPublic {
            username: None,
            user_type: None,
            appearance: Box::new(appearance),
        }
    }
}