# console-chat/openapi/src/models2/appearance_public.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct AppearancePublic {
    #[serde(rename = "color")]
    pub color: String,
}

impl AppearancePublic {
    pub fn new(color: String) -> AppearancePublic {
        AppearancePublic {
            color,
        }
    }
}