# console-chat/openapi/src/models2/key_response.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct KeyResponse {
    #[serde(rename = "encrypted_symmetric_key")]
    pub encrypted_symmetric_key: String,
    #[serde(rename = "check_msg")]
    pub check_msg: String,
    #[serde(rename = "sender_public_key")]
    pub sender_public_key: String,
}

impl KeyResponse {
    pub fn new(encrypted_symmetric_key: String, check_msg: String, sender_public_key: String) -> KeyResponse {
        KeyResponse {
            encrypted_symmetric_key,
            check_msg,
            sender_public_key,
        }
    }
}