# console-chat/openapi/src/models2/message_send.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct MessageSend {
    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
    pub r#type: Option<models::MessageType>,

    #[serde(
        rename = "content",
        default,
        with = "::serde_with::rust::double_option",
        skip_serializing_if = "Option::is_none"
    )]
    pub content: Option<Option<Box<models::Content>>>,

    #[serde(rename = "send_at", skip_serializing_if = "Option::is_none")]
    pub send_at: Option<String>,

    #[serde(
        rename = "data",
        default,
        with = "::serde_with::rust::double_option",
        skip_serializing_if = "Option::is_none"
    )]
    pub data: Option<Option<Box<serde_json::Value>>>,
}

impl MessageSend {
    pub fn new() -> MessageSend {
        MessageSend {
            r#type: None,
            content: None,
            send_at: None,
            data: None,
        }
    }
}