# console-chat/openapi/src/models2/online_response.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct OnlineResponse {
    #[serde(rename = "token")]
    pub token: Box<models::Token>,
    #[serde(rename = "user")]
    pub user: uuid::Uuid,
}

impl OnlineResponse {
    pub fn new(token: models::Token, user: uuid::Uuid) -> OnlineResponse {
        OnlineResponse {
            token: Box::new(token),
            user,
        }
    }
}