# console-chat/openapi/src/models2/plaintext.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct Plaintext {
    #[serde(rename = "content")]
    pub content: String,
}

impl Plaintext {
    pub fn new(content: String) -> Plaintext {
        Plaintext {
            content,
        }
    }
}