# console-chat/openapi/src/models2/system_message.rs
/*
 * Console Chat API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct SystemMessage {
    #[serde(rename = "content")]
    pub content: String,
    #[serde(rename = "online_users")]
    pub online_users: i32,
}

impl SystemMessage {
    pub fn new(content: String, online_users: i32) -> SystemMessage {
        SystemMessage {
            content,
            online_users,
        }
    }
}